Délègue à un helper les requêtes sur les événements d'une restitution
